backend/requirements.txt
flask~=3.1.0
PyJWT~=2.9.0
flask-jwt-extended~=4.7.1
python-dotenv~=1.0.1
requests~=2.32.3